+++ jlegmed-core/src/main/java/io/jexxa/jlegmed/plugins/persistence/repository/Repository.java
@@ -23,12 +23,21 @@ public class Repository<T, K> {
         this.keyFunction = keyFunction;
     }
 
+    Repository(Class<T> aggregateClazz,
+               Function<T,K> keyFunction,
+               String storageName,
+               FilterContext filterContext)
+    {
+        this.iRepository = createRepository(aggregateClazz, keyFunction, storageName, filterContext.properties());
+        this.keyFunction = keyFunction;
+    }
+
     /**
      * Updates the given aggregate inside the repository.
      * @param aggregate that should be updated
      * @pre Given aggregate must be added by using {@link #add(Object)}}
      */
-    public T update(T aggregate)
+    synchronized public T update(T aggregate)
     {
         iRepository.update(aggregate);
         return aggregate;
@@ -39,7 +48,7 @@ public class Repository<T, K> {
      * @param key to the aggregate to be removed
      * @pre Aggregate must be added by using {@link #add(Object)}}
      */
-    public K remove(K key) {
+    synchronized public K remove(K key) {
         iRepository.remove(key);
         return key;
     }
@@ -49,7 +58,7 @@ public class Repository<T, K> {
      * @param aggregate that should be added
      * @pre Aggregate must not be added before
      */
-    public T add(T aggregate) {
+    synchronized public T add(T aggregate) {
         iRepository.add(aggregate);
         return aggregate;
     }
@@ -58,7 +67,7 @@ public class Repository<T, K> {
      * Either updates an aggregate if exists, or adds an aggregate to this repository
      * @param aggregate that should be added or updated
      */
-    public T put(T aggregate) {
+    synchronized public T put(T aggregate) {
         if ( get(keyFunction.apply(aggregate)).isPresent()) {
             return update(aggregate);
         }
@@ -71,15 +80,18 @@ public class Repository<T, K> {
      * @param key that identifies the aggregate
      * @return Optional of aggregate. Optional is empty if given key finds no aggregate.
      */
-    public Optional<T> get(K key)
+    synchronized public Optional<T> get(K key)
     {
         return iRepository.get(key);
     }
 
-    public List<T> get()
+    synchronized public List<T> get()
     {
 
         return iRepository.get();
     }
 
+    protected  IRepository<T, K> repository() {
+        return iRepository;
+    }
 }